+++ packages/polythene-core-list/src/list.js
@@ -7,7 +7,49 @@ export const getElement = vnode =>
 
 export const theme = customTheme;
 
+const onSelect = (event, vnode) => {
+  const state = vnode.state;
+  const attrs = vnode.attrs;
+  if (attrs.onSelect) {
+    const highlightIndex = state.highlightIndex();
+    const data = {
+      event,
+      index: highlightIndex,
+      dom: state.tiles[highlightIndex].dom,
+      attrs: state.tiles[highlightIndex].attrs
+    };
+    attrs.onSelect(data);
+  }
+};
+
+export const getInitialState = (vnode, createStream) => {
+  const attrs = vnode.attrs;
+  const highlightIndex = createStream(attrs.defaultHighlightIndex !== undefined
+    ? attrs.defaultHighlightIndex
+    : -1);
+  const registerTile = state => (index, data) => state.tiles[index] = data;
+  return {
+    tiles: [],
+    highlightIndex,
+    registerTile,
+    redrawOnUpdate: createStream.merge([highlightIndex])
+  };
+};
+
+export const onMount = vnode => {
+  const state = vnode.state;
+  const attrs = vnode.attrs;
+  if (attrs.keyboardControl) {
+    state.highlightIndex.map(index => {
+      if (state.tiles[index]) {
+        state.tiles[index].dom.focus();
+      }
+    });
+  }
+};
+
 export const createProps = (vnode, { keys: k }) => {
+  const state = vnode.state;
   const attrs = vnode.attrs;
   return Object.assign(
     {},
@@ -24,11 +66,31 @@ export const createProps = (vnode, { keys: k }) => {
         attrs.tone === "light" ? "pe-light-tone" : null,
         attrs.className || attrs[k.class],
       ].join(" "),
+    },
+    attrs.keyboardControl && {
+      [k.onkeydown]: e => {
+        const highlightIndex = state.highlightIndex();
+        if (e.key === "ArrowDown" || e.key === "ArrowRight") {
+          e.preventDefault(); // prevent scrolling the page
+          const newIndex = Math.min(state.tiles.length - 1, highlightIndex + 1);
+          state.tiles[newIndex].dom.focus();
+        } else if (e.key === "ArrowUp" || e.key === "ArrowLeft") {
+          e.preventDefault(); // prevent scrolling the page
+          const newIndex = Math.max(0, highlightIndex - 1);
+          state.tiles[newIndex].dom.focus();
+        } else if (e.key === "Enter") {
+          onSelect(e, vnode);
+        } else if (e.key === "Escape") {
+          state.tiles[highlightIndex].dom.blur();
+          state.highlightIndex(-1);
+        }
+      },
     }
   );
 };
 
 export const createContent = (vnode, { renderer: h, requiresKeys, keys: k, ListTile }) => {
+  const state = vnode.state;
   const attrs = vnode.attrs;
   let headerOpts;
   if (attrs.header) {
@@ -41,16 +103,48 @@ export const createContent = (vnode, { renderer: h, requiresKeys, keys: k, ListT
       headerOpts[k.class] || null
     ].join(" ");
   }
+  const highlightIndex = state.highlightIndex();
+  const tiles = attrs.tiles
+    ? attrs.tiles
+    : attrs.content
+      ? attrs.content
+      : attrs.children || vnode.children;
+  let index = -1;
   return [
     headerOpts ? h(ListTile, Object.assign(
       {},
       requiresKeys ? { key: "header" } : null,
-      headerOpts
+      headerOpts,
+      {
+        header: true
+      }
     )) : null,
-    attrs.tiles
-      ? attrs.tiles
-      : attrs.content
-        ? attrs.content
-        : attrs.children || vnode.children
+    attrs.keyboardControl
+      ? tiles.map(tileOpts => {
+        if (!tileOpts.header) {
+          index++;
+        }
+        return tileOpts.tag !== undefined
+          ? tileOpts
+          : h(ListTile, Object.assign(
+            {},
+            tileOpts,
+            !tileOpts.header && {
+              keyboardControl: true,
+              register: state.registerTile(state),
+              setHighlightIndex: state.highlightIndex,
+              index,
+              defaultHighlight: highlightIndex === index,
+              events: Object.assign(
+                {},
+                tileOpts.events,
+                {
+                  [k.onclick]: e => onSelect(e, vnode)
+                }
+              )
+            }
+          ));
+      })
+      : tiles
   ];
 };