+++ packages/polythene-mithril-list-tile/src/index.js
@@ -1,9 +1,9 @@
-import { ViewComponent } from "polythene-mithril-base";
+import { StateComponent } from "polythene-mithril-base";
 import { coreListTile as core } from "polythene-core-list-tile";
 import { Icon } from "polythene-mithril-icon";
 import { Ripple } from "polythene-mithril-ripple";
 
-export const ListTile = ViewComponent(Object.assign(
+export const ListTile = StateComponent(Object.assign(
   {},
   core,
   {